Clean up old search page
