+++ libs/common/src/index.ts
@@ -19,7 +19,6 @@ export * from './lib/page-not-found';
 export * from './lib/pagination';
 export * from './lib/pipes';
 export * from './lib/resources';
-export * from './lib/search-widget/search';
 export * from './lib/search-widget';
 export * from './lib/select-account-kb';
 export * from './lib/services';
