+++ libs/common/src/lib/search-widget/index.ts
@@ -1,4 +1,3 @@
-export * from './search';
 export * from './widgets';
 export * from './search-widget.models';
