--- libs/common/src/lib/search-widget/search/index.ts
@@ -1,2 +1,0 @@
-export * from './deprecated-search.component';
-export * from './search.module';
